xml/parsing/manakai/peref-3.dat

#data
<!DOCTYPE a SYSTEM "http://a/"><a x="&fuga;d"/>
#resource http://a/
<!ENTITY % foo SYSTEM "http://b/">
<!ENTITY %foo; "bcc">
#resource http://b/
fuga
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]2;1;xml:dtd:ext decl;;;w
[2]1;1;no XML decl;;;s
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
|   x="bccd"

#data
<!DOCTYPE a SYSTEM "http://a/"><a x="&fuga;d"/>
#resource http://a/
<!ENTITY % foo SYSTEM "http://b/">
<!ENTITY %foo;fuga "bcc">
#resource http://b/
<?xml encoding="utf-8"?>
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
|   x="bccd"

#data
<!DOCTYPE a SYSTEM "http://a/"><a x="&fuga;d"/>
#resource http://a/
<!ENTITY % foo SYSTEM "http://b/">
<!ENTITY %foo; "bcc">
#resource http://b/
<?xml encoding="utf-8"?>fuga
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
|   x="bccd"

#data
<!DOCTYPE a SYSTEM "http://a/"><a x="&fuga;d"/>
#resource http://a/
<!ENTITY % foo SYSTEM "http://b/">
<!ENTITY %foo; "bcc">
#resource http://b/
<?xml encoding="utf-8"?>
fuga
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]2;1;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
|   x="bccd"

#data
<!DOCTYPE a SYSTEM "http://a/"><a x="&fuga;d"/>
#resource http://a/
<!ENTITY % foo SYSTEM "http://b/">
<!ENTITY %foo; "bcc">
#resource http://b/
<?hogfe?>fuga
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[2]1;1;no XML decl;;;s
[2]1;8;tag in pe in markup declaration;;;m
1;38;entity not declared;;&fuga;;m
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
|   x="&fuga;d"
#entities
| <!ENTITY % foo "" "" "http://b/" >

#data
<!DOCTYPE a SYSTEM "http://a/"><a x="&fuga;d"/>
#resource http://a/
<!ENTITY % foo SYSTEM "http://b/">
<!ENTITY %foo; "bcc">
#resource http://b/
<?hogfe?>
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[2]1;1;no XML decl;;;s
[2]1;8;tag in pe in markup declaration;;;m
1;38;entity not declared;;&fuga;;m
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
|   x="&fuga;d"

#data
<!DOCTYPE a SYSTEM "http://a/"><a x="&fuga;d"/>
#resource http://a/
<!ENTITY % foo SYSTEM "http://b/">
<!ENTITY %foo; "bcc">
#resource http://b/
<?hogfe?
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[2]1;1;no XML decl;;;s
[2]1;7;tag in pe in markup declaration;;;m
1;38;entity not declared;;&fuga;;m
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
|   x="&fuga;d"

#data
<!DOCTYPE a SYSTEM "http://a/"><a x="&fuga;d"/>
#resource http://a/
<!ENTITY % foo SYSTEM "http://b/">
<!ENTITY %foo; "bcc">
#resource http://b/
<?hogfe
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[2]1;1;no XML decl;;;s
[2]1;6;tag in pe in markup declaration;;;m
1;38;entity not declared;;&fuga;;m
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
|   x="&fuga;d"

#data
<!DOCTYPE a SYSTEM "http://a/"><a x="&fuga;d"/>
#resource http://a/
<!ENTITY % foo SYSTEM "http://b/">
<!ENTITY %foo; "bcc">
#resource http://b/
<?hogfe  
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[2]1;1;no XML decl;;;s
[2]1;8;tag in pe in markup declaration;;;m
1;38;entity not declared;;&fuga;;m
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
|   x="&fuga;d"

#data
<!DOCTYPE a SYSTEM "http://a/"><a x="&fuga;d"/>
#resource http://a/
<!ENTITY % foo SYSTEM "http://b/">
<!ENTITY %foo; "bcc">
#resource http://b/
<?hogfe???
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[2]1;1;no XML decl;;;s
[2]1;9;tag in pe in markup declaration;;;m
1;38;entity not declared;;&fuga;;m
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
|   x="&fuga;d"

#data
<!DOCTYPE a SYSTEM "http://a/"><a x="&fuga;d"/>
#resource http://a/
<!ENTITY % foo SYSTEM "http://b/">
<!ENTITY %foo; "bcc">
#resource http://b/
<?
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[2]1;1;no XML decl;;;s
[2]1;1;tag in pe in markup declaration;;;m
1;38;entity not declared;;&fuga;;m
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
|   x="&fuga;d"

#data
<!DOCTYPE a SYSTEM "http://a/"><a x="&fuga;d"/>
#resource http://a/
<!ENTITY % foo SYSTEM "http://b/">
<!ENTITY %foo; "bcc">
#resource http://b/
<
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[2]1;1;no XML decl;;;s
[2]1;3;tag in pe in markup declaration;;;m
1;38;entity not declared;;&fuga;;m
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
|   x="&fuga;d"
#entities
| <!ENTITY % foo "" "" "http://b/" >

#data
<!DOCTYPE a SYSTEM "http://a/"><a x="&fuga;d"/>
#resource http://a/
<!ENTITY % foo SYSTEM "http://b/">
<!ENTITY %foo; "bcc">
#resource http://b/
  <?xml encoding="utf-8"?>
fuga
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[1]2;1;xml:dtd:ext decl;;;w
[2]1;1;no XML decl;;;s
[2]1;9;xml:dtd:string after name;;;m
[2]1;26;mdc in pe in md;;;m
1;38;entity not declared;;&fuga;;m
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
|   x="&fuga;d"

#data
<!DOCTYPE a SYSTEM "http://a/"><a x="&fuga;d"/>
#resource http://a/
<!ENTITY % foo SYSTEM "http://b/">
<!ENTITY %foo; "bcc">
#resource http://b/
  <?xml?>
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[2]1;1;no XML decl;;;s
[2]1;9;mdc in pe in md;;;m
[1]2;1;xml:dtd:ext decl;;;w
1;38;entity not declared;;&fuga;;m
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
|   x="&fuga;d"
#entities
| <!ENTITY <?xml? "" "" "" >
| <!ENTITY % foo "" "" "http://b/" >

#data
<!DOCTYPE a SYSTEM "http://a/"><a x="&fuga;d"/>
#resource http://a/
<!ENTITY % foo SYSTEM "http://b/">
<!ENTITY %foo; "bcc">
#resource http://b/
<? aaa>hoge
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[2]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[2]1;6;tag in pe in markup declaration;;;m
1;38;entity not declared;;&fuga;;m
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
|   x="&fuga;d"

#data
<!DOCTYPE a SYSTEM "http://a/"><a x="&fuga;d"/>
#resource http://a/
<!ENTITY % foo SYSTEM "http://b/">
<!ENTITY %foo; "bcc">
#resource http://b/
<<hoge
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[2]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[2]1;1;tag in pe in markup declaration;;;m
[2]1;2;tag in pe in markup declaration;;;m
1;38;entity not declared;;&fuga;;m
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
|   x="&fuga;d"

#data
<!DOCTYPE a SYSTEM "http://a/"><a x="&fuga;d"/>
#resource http://a/
<!ENTITY % foo SYSTEM "http://b/">
<!ENTITY %foo; "bcc">
#resource http://b/
<!hoge
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[2]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[2]1;1;tag in pe in markup declaration;;;m
[2]1;2;tag in pe in markup declaration;;;m
1;38;entity not declared;;&fuga;;m
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
|   x="&fuga;d"

#data
<!DOCTYPE a SYSTEM "http://a/"><a x="&fuga;d"/>
#resource http://a/
<!ENTITY % foo SYSTEM "http://b/">
<!ENTITY %foo; "bcc">
#resource http://b/ escaped
<\u0000hoge
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[2]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[2]1;1;tag in pe in markup declaration;;;m
[2]1;2;NULL;;;m
1;38;entity not declared;;&fuga;;m
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
|   x="&fuga;d"

#data
<!DOCTYPE a SYSTEM "http://a/"><a x="&fuga;d"/>
#resource http://a/
<!ENTITY % foo SYSTEM "http://b/">
<!ENTITY %foo; "bcc">
#resource http://b/ escaped
<%hoge
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[2]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[2]1;1;tag in pe in markup declaration;;;m
[2]1;2;tag in pe in markup declaration;;;m
1;38;entity not declared;;&fuga;;m
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
|   x="&fuga;d"

#data
<!DOCTYPE a SYSTEM "http://a/"><a x="&fuga;d"/>
#resource http://a/
<!ENTITY % foo SYSTEM "http://b/">
<!ENTITY %foo; "bcc">
#resource http://b/ escaped
<&hoge
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[2]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[2]1;1;tag in pe in markup declaration;;;m
[2]1;2;tag in pe in markup declaration;;;m
1;38;entity not declared;;&fuga;;m
#document
| <!DOCTYPE a "" "http://a/">
| <{} a>
|   x="&fuga;d"
